input: add support for Chipone ICNL9916 based touchscreens
The Chipone ICNL9916 is a combined display driver and touchscreen IC
found in various mobile phones. Add a driver for the touchscreen part.

This driver is based on the existing drivers for the distantly related
ICN8318 and ICN8505 controllers. ICNL9916 uses a completely different
command-based protocol via I2C, but still has some similarities with
them.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>