power: supply: sc27xx: automatically calibrate on every boot
The last saved capacity value may become outdated if the device is
powered off for a long time or booted in a state in which the fuel
gauge driver is not active. Instead of recomputing the capacity only
after a full shutdown, which happens very rarely on devices with a
non-removable battery, adjust it on every boot. Try to use a value
close to the saved one to avoid large jumps due to measurement
inaccuracies.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>