ASoC: sprd: do not keep reference to DMA parameters
The DMA parameters may be freed after hw_free is called, so it cannot
be assumed that the pointer will remain valid. Since the parameters are
not used outside hw_params, just pass them directly to
sprd_pcm_request_dma_channel instead of storing them as part of the
driver's private state.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>