ASoC: sprd: add support for single-channel DMA with interleaved data
Advertise both INTERLEAVED and NONINTERLEAVED in the PCM driver and use
the selected access type to determine the effective number of DMA
channels. This makes the CPU DAI driver responsible for selecting the
right access type using snd_pcm_hw_constraint_mask64(), and fixes the
problem that the PCM driver currently advertises only the INTERLEAVED
access type but then splits the data between multiple DMA channels as
if it were NONINTERLEAVED.

When MCDT is used, there is only a single FIFO for two audio channels,
so the INTERLEAVED format is correct. However, it should then be sent to
the FIFO through one DMA channel. If each audio channel has a separate
destination, which may be the case if MCDT and the DSP are bypassed, the
NONINTERLEAVED format should be used instead.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>