gnss: add driver for Unisoc SoCs
Add a driver for integrated GNSS receivers found in Unisoc SoCs.

These devices encapsulate NMEA sequences within a custom command-based
protocol. This protocol is currently decoded in the driver and not
exposed to userspace, although this may be necessary later to allow
configuring the device.

For now, the driver just provides a ready-to-use NMEA stream and uses
commands for power management.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>